플레이어 구조체 파일 추가. 기본 로직은 게임 구조체로 옮김. 플레이어 이름 5자 이상은 무시, 마지막 줄에 이름 출력.
